nitroducing a new set of classes and structures to manage configuration files stored in a xml-like format.
The format is a subset of xml, it is not meant to be used to interoperate with
other applications, but it's a good way to handle configuration files
inside a single app.


Entry class (used in Linklist) have been modified to let the user
attach a void * pointer to use as the entry value. get_value() and
set_value() methods have been implemented to let the user manage this pointer.


git-svn-id: svn://dyne.org/rastasoft/muse@444 383723c8-4afa-0310-b8a8-b1afb83214fc